refactor (#75): refactor api endpoint and fix some bug
