+++ src/modules/users/user.service.ts
@@ -3,14 +3,20 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
+import { TOKEN_MESSAGES, USERS_MESSAGES } from '@src/constants/message';
 import { generateHash, validateHash } from '@src/common/utils';
 
 import { ChangePasswordDto } from './dto/change-password.dto';
+import { ConfigService } from '@nestjs/config';
+import { ForgotPasswordDto } from './dto/forgot-password.dto';
+import { JwtService } from '@nestjs/jwt';
+import { MailerService } from '@nestjs-modules/mailer';
 import { PageDto } from '@src/common/dto/page.dto';
 import { PageMetaDto } from '@src/common/dto/page-meta.dto';
 import { PrismaService } from '@src/shared/prisma/prisma.service';
+import { ResetPasswordDto } from './dto/reset-password.dto';
 import { StorageService } from '@src/shared/storage/services/storage.service';
-import { USERS_MESSAGES } from '@src/constants/message';
+import { TokenInvalidException } from '@src/exceptions';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from '@prisma/client';
 import { UserEntity } from './entities/user.entity';
@@ -22,6 +28,9 @@ import { v4 as uuidv4 } from 'uuid';
 export class UserService {
   constructor(
     private prisma: PrismaService,
+    private readonly config: ConfigService,
+    private readonly jwtService: JwtService,
+    private readonly mailerService: MailerService,
     private readonly storageService: StorageService,
   ) {}
 
@@ -62,9 +71,168 @@ export class UserService {
       where: { id },
     });
 
-    if (!user) throw new NotFoundException(`User #${id} not found`);
+    if (!user) throw new NotFoundException(USERS_MESSAGES.USER_NOT_FOUND);
 
     return user;
+  }
+
+  private signForgotPasswordToken({ userId }: { userId: number }) {
+    return this.jwtService.sign(
+      { userId },
+      {
+        secret: this.config.getOrThrow<string>('auth.jwtForgotPasswordSecret'),
+        expiresIn: `${this.config.getOrThrow<string>(
+          'auth.jwtForgotPasswordExpires',
+        )}s`,
+      },
+    );
+  }
+
+  private sendForgotPasswordMail({
+    email,
+    token,
+    name,
+  }: {
+    email: string;
+    token: string;
+    name: string;
+  }) {
+    const apiPrefix =
+      this.config.getOrThrow<string>('app.apiPrefix') +
+      '/' +
+      this.config.getOrThrow<string>('app.apiVersion');
+
+    const resetLink = `${this.config.getOrThrow(
+      'app.appURL',
+    )}/${apiPrefix}/auth/verify-forgot-password?token=${token}`;
+
+    return this.mailerService.sendMail({
+      to: email,
+      from: 'elearningapp@gmail.com',
+      subject: 'Email reset forgot password for leaning app',
+      template: './forgot-password',
+      context: {
+        name,
+        resetLink: resetLink,
+      },
+    });
+  }
+
+  async forgotPassword({ email }: ForgotPasswordDto) {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        email,
+      },
+    });
+
+    if (!user) throw new NotFoundException(USERS_MESSAGES.USER_NOT_FOUND);
+
+    const token = this.signForgotPasswordToken({
+      userId: user.id,
+    });
+
+    await this.prisma.user.update({
+      where: {
+        email,
+      },
+      data: {
+        forgotPasswordToken: token,
+      },
+    });
+
+    const name = user.firstName + user.lastName;
+
+    await this.sendForgotPasswordMail({ email, token, name });
+
+    return {
+      message: USERS_MESSAGES.CHECK_EMAIL_TO_RESET_PASSWORD,
+    };
+  }
+
+  verifyForgotPassword(token: string) {
+    const frontendURL = this.config.get('app.frontendURL');
+    return `${frontendURL}/reset-password?token=${token}`;
+  }
+
+  async resetPassword(resetPasswordDto: ResetPasswordDto) {
+    try {
+      const { token, password } = resetPasswordDto;
+      const payload: {
+        userId: number;
+      } = this.jwtService.verify(token, {
+        secret: this.config.getOrThrow<string>('auth.jwtForgotPasswordSecret'),
+        ignoreExpiration: false,
+      });
+
+      const user = await this.prisma.user.findUnique({
+        where: {
+          id: payload.userId,
+        },
+      });
+
+      if (!user) throw new NotFoundException(USERS_MESSAGES.USER_NOT_FOUND);
+
+      if (user.forgotPasswordToken != token)
+        throw new BadRequestException(
+          USERS_MESSAGES.FORGOT_PASSWORD_TOKEN_INVALID,
+        );
+
+      await this.prisma.user.update({
+        where: {
+          email: user.email,
+        },
+        data: {
+          password: generateHash(password),
+          forgotPasswordToken: null,
+        },
+      });
+
+      return {
+        message: USERS_MESSAGES.RESET_PASSWORD_SUCCESSFUL,
+      };
+    } catch (error) {
+      throw new TokenInvalidException(TOKEN_MESSAGES.TOKEN_IS_INVALID);
+    }
+  }
+
+  /**
+   * Changes the password of a user.
+   * @param id - The ID of the user whose password needs to be changed.
+   * @param changePasswordDto - An object containing the old password, new password, and confirm password.
+   * @returns An object with a success message.
+   * @throws NotFoundException if the user is not found or the passwords do not match.
+   */
+  async changePassword(id: number, changePasswordDto: ChangePasswordDto) {
+    const { newPassword, oldPassword, confirmPassword } = changePasswordDto;
+
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException(USERS_MESSAGES.USER_NOT_FOUND);
+    }
+
+    if (newPassword !== confirmPassword) {
+      throw new BadRequestException(USERS_MESSAGES.PASSWORD_NOT_MATCH);
+    }
+
+    if (!(await validateHash(oldPassword, user.password))) {
+      throw new BadRequestException(USERS_MESSAGES.PASSWORD_NOT_MATCH);
+    }
+
+    const hashedNewPassword = generateHash(newPassword);
+
+    await this.prisma.user.update({
+      where: { id },
+      data: {
+        password: hashedNewPassword,
+      },
+    });
+
+    return {
+      message: USERS_MESSAGES.CHANGE_PASSWORD_SUCCESSFULLY,
+    };
   }
 
   async update(
@@ -106,46 +274,6 @@ export class UserService {
         'role',
         'status',
       ]),
-    };
-  }
-
-  /**
-   * Changes the password of a user.
-   * @param id - The ID of the user whose password needs to be changed.
-   * @param changePasswordDto - An object containing the old password, new password, and confirm password.
-   * @returns An object with a success message.
-   * @throws NotFoundException if the user is not found or the passwords do not match.
-   */
-  async changePassword(id: number, changePasswordDto: ChangePasswordDto) {
-    const { newPassword, oldPassword, confirmPassword } = changePasswordDto;
-
-    const user = await this.prisma.user.findUnique({
-      where: { id },
-    });
-
-    if (!user) {
-      throw new NotFoundException(USERS_MESSAGES.USER_NOT_FOUND);
-    }
-
-    if (newPassword !== confirmPassword) {
-      throw new BadRequestException(USERS_MESSAGES.PASSWORD_NOT_MATCH);
-    }
-
-    if (!(await validateHash(oldPassword, user.password))) {
-      throw new BadRequestException(USERS_MESSAGES.PASSWORD_NOT_MATCH);
-    }
-
-    const hashedNewPassword = generateHash(newPassword);
-
-    await this.prisma.user.update({
-      where: { id },
-      data: {
-        password: hashedNewPassword,
-      },
-    });
-
-    return {
-      message: USERS_MESSAGES.CHANGE_PASSWORD_SUCCESSFULLY,
     };
   }
